Refactor files.
